feat: JS build support (#3548)
## Relevant issue(s)

Resolves #3547

## Description

This PR adds JS build support via the `node` package`. 

It also adds a new test action to ensure JS builds are not broken.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Added JS specific test.

Specify the platform(s) on which this was tested:
- MacOS