test(i): Wait for database to close before progressing (#3593)
## Relevant issue(s)

Resolves #3592

## Description

Wait for the database to close before progressing to the next test.

Previously the database would close in the background while the next
test began, this created a race condition when setting `os.Stderr`.

Credit to Chris, as he suffered with this bug for a while and was a bug
help in figuring out where it was.