Move sequence to sub-package
Hosting it in db makes it impossible to use anywhere else, including in db sub-packages (without really nasty workarounds).

Note: This is not actually required for this PR, as I was able to completely remove Collection.ID - I will however need this in later PRs as Collection.Root and field ids will need to make use of this.