fix(i): Node identity as default instead of fallback signer (#3629)
## Relevant issue(s)

Resolves #3628 

## Description

The PR revert the inclusion of the fallback signer configuration. There
was a bug where if the request did not provide an identity, then the
node identity would be used instead of the fallback signer. This was an
opaque bug because in its current state, the fallback signer is always
the node identity.

At standup we concluded that it would be best to just keep the node
identity as the default at all times for now.