refactor(i): Remove the document acp instance from node (#3670)
## Relevant issue(s)
Resolves #3669 

## Description
This was introduced in the purging acp state work done here:
https://github.com/sourcenetwork/defradb/pull/3363
There is no need to have a duplicated acp instance on the node and on
the DB instance.
Later there was a PR that started passing/depending on this instance by
passing into the p2p node start function.