refactor: Rename CollectionDescription to CollectionVersion (#3676)
## Relevant issue(s)

Resolves #3672

## Description

Renames CollectionDescription to CollectionVersion.

I used a handful of find and replaces, plus small manual tweaks for
this.

There are still variables etc that contain some variant of 'desc' in
their names. I do not think it is worth being too thorough atm as this
code space remains in a state of flux, with some notable stuff such as
CollectionDefinition and SchemaDescription yet to be removed - so I ask
reviewers to give a very quick scan to see if anything important has
been missed, and to please avoid raising lesser misses here.