Replace commit fieldID with fieldName
Whilst fieldName is less efficient (for now) when providing anything but the composite identifier (the most common/important case), it is much simpler for users.  FieldID is soon to become an internal-only concept, and will make even less sense to users than it currently does.

Using `_C` to identify composites is questionable, IMO, it is good enough for now, and lets me crack on with my primary goal.  Long term we might want to split this out to another param, e.g. '_compositesOnly'.