refactor: Allow retrieval of txn from ctx everywhere (#3705)
## Relevant issue(s)

Resolves #3704

## Description

Allows the retrieval of txn from ctx everywhere. Does not remove the
many function params that could/should be removed because of this change
(we can do that gradually over time).

Is annoying and messy to have to pass both independently into most
functions in our codebase, especially when one lives within the other.

Broken out of #3699