feat: JavaScript event system (#3794)
## Relevant issue(s)

Resolves #3793

## Description

This PR adds a JavaScript based event system. This is a pre-requisite
for the JavaScript p2p system.

Note: Some of the events contain channels and will not be compatible.
This will be fixed in a future PR.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Added unit tests.

Specify the platform(s) on which this was tested:
- MacOS