refactor: ACP Identity (#3808)
## Relevant issue(s)

Resolves #3809

## Description

This PR focuses on refactoring the ACP identity from a single struct to
multiple interfaces with concrete implementations. All code and tests
now use the new interfaces, making identity code easier to maintain and
extend.

- Added multiple identity interfaces and concrete implementations
- Added identity factory
- Updated all usages to use new interfaces
- Refactored and improved related code and tests

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly (where relevant).
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

- All unit and integration tests updated and passing
- Manual testing
- Lint and formatting checks

Tested on:
- MacOS

**NOTE:**
- This is a big refactor. You'll probably need to update how you use
identities due to breaking changes.