refactor: Simplify interfaces usage (#3739)
## Relevant issue(s)

Resolves #3738
Resolves #3572 

## Description
This is Part 1 of the interface refactor. 

This PR tries to localize the importance of interfaces and makes a step
towards trimming the number of methods on the interfaces that are used
globally.

Key changes:
- The `Multistore` interface has been completely removed with functions
wrapping the root store with sub-store prefixes have been added. This
allows any object that has a reference to the rootstore to reach the
appropriate sub-store.
- The replicator and p2p collection code has been moved back to the
`net` package so that the `client.P2P` interface can be truly enforcing
the minimum implementation (ensures `net` is a P2P implementation).
- The `client.DB` interface now only contains methods that are the
minimum required by our official client implementations (CLI, HTTP,
Node).

Coming up next:
- Implementation of `client.DB` and `client.P2P` by `node.Node` itself.
- [x] Potentially more work to centralize transactions in `datastore`
and further reduce the `Txn` interface.
- Moving `db` out of `internal`