feat: Add WASM support for Playground (#3813)
## Relevant issue(s)

Resolves [#3812](https://github.com/sourcenetwork/defradb/issues/3812)

## Description

This PR adds WASM support to the playground. Now there are two modes:
**wasm** (runs in browser) and **remote** (connects to DefraDB node).

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [ ] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Tested in browser with common queries like `addSchema`, `patchSchema`,
`addView`, and `execRequest`.

Specify the platform(s) on which this was tested:
- MacOS / Chrome browser