chore: Add TOC to playground readme and TOC check CI job (#3827)
## Relevant issue(s)

Resolves [#3826](https://github.com/sourcenetwork/defradb/issues/3826)

## Description

This PR adds TOC to the playground readme and the corresponding CI job
to run the TOC check.

## Tasks

- [ ] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [ ] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

(*replace*) Describe the tests performed to verify the changes. Provide
instructions to reproduce them.

Specify the platform(s) on which this was tested:
- MacOS