feat: SourceHub ACP JS (#3863)
## Relevant issue(s)

Resolves [#3862](https://github.com/sourcenetwork/defradb/issues/3862)

## Description

This PR implements SourceHub ACP for JS client (`sourcehub_acp_js.go`),
adds `secp256r1` support, and introduces multiple improvements to the JS
defradb client.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

(*replace*) Describe the tests performed to verify the changes. Provide
instructions to reproduce them.

Specify the platform(s) on which this was tested:
- Debian Linux
- MacOS