feat: Provide C Bindings (#3787)
## Relevant issue(s)

Resolves #3362

## Description

There are several things occurring in this PR. 

**- C Bindings:**

There exists a directory, cbindings, which contains various .go files,
as well as a `defra_structs.h` file. These files implement the
functionality that will be exported to the shared object. In other
words, these are the functions that will be made available to the users
in C. Currently, they are grouped in a similar to way as the subcommands
in the Defra CLI, although there has already been some feedback
regarding whether or not that should be the case. These functions will
likely be re-grouped, re-named, or both.

To allow for the building, a new section of the makefile has been added,
`build-c-shared-linux` which is where current development is focused.
Building for Windows and/or Android are not currently included in this
PR.

**- Integration Test Support:**

A new directory exists called `tests/integration/clients/c`. This
contains code that wraps a large portion of the C binding code for use
in our integration tests. There also exists a number of files which
contain functions duplicating the behavior of those found in the
`cbindings` directory. The reason for this, is that we have a layered
problem:

1) The integration tests require that we be able to pass complex objects
(such as a Node), which is not possible through calling into a .so
2a) We can, as a workaround, call the functions directly using C-Types,
to test the logic...
2b) ...but to be exportable, the functions must exist in a package named
"main", which we cannot import and use...
2c) ...and additionally, due to the way Go handles C Types, we cannot
pass a C structure from one package to another.

Therefore, it's a little bit messy, but we have the exported functions,
and the appropriate .h file, all in the `tests/integration/clients/c`
folder. This allows the integration tests to run. Hopefully, we can find
a better approach, for the long-term, moving forward.

**- Key Takeaway / Where we are at:**

Testing indicates that the C functionality is working well. However,
properly incorporating a C client into the integration test suite that
we have is a bit hacky. This PR attempts to create a base to work from,
where the core logic works correctly, and is testable.

**- Important Note on P2P Functionality:**

Currently, there exist wrapper functions for P2P related functionality
(i.e. P2PsetReplicator). However, network tests are skipped. The
architected design of the C bindings relies on a singular globalNode
which is accessed and used by all of the exported functions. Even if two
client wrappers were instantiated, they would be utilizing the same node
which is problematic.

*This is the biggest open issue to be addressed. But as per the team
discussion, this will be addressed in a separate PR*

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL