fix: Playground keypair reset (#3922)
## Relevant issue(s)

Resolves #3921

## Description

This PR implements a fix to reset the keypair automatically when using
local sourcehub deployment.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

- Verified that the keypair is re-generated on every page reload when
using local sourcehub deployment.

Specify the platform(s) on which this was tested:
- MacOS (Chrome)